+++ api/authentication/authentication_test.go
@@ -41,16 +41,55 @@ type GetPasswordSafeAuthenticationConfig struct {
 // the default API version 3.0 will be used
 var apiVersion string = "3.1"
 
+<<<<<<< HEAD
 func TestSignOut(t *testing.T) {
+=======
+var authParamsOauth *AuthenticationParametersObj
+var authParamsKey *AuthenticationParametersObj
+var zapLogger *logging.ZapLogger
+
+func InitializeGlobalConfig() {
+
+>>>>>>> main
 	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	zapLogger = logging.NewZapLogger(logger)
 
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
+
+	// authentication using Oauth Method
+	authParamsOauth = &AuthenticationParametersObj{
+		HTTPClient:                 *httpClientObj,
+		BackoffDefinition:          backoffDefinition,
+		EndpointURL:                "https://fake.api.com:443/BeyondTrust/api/public/v3/",
+		APIVersion:                 apiVersion,
+		ClientID:                   "fakeone_a654+9sdf7+8we4f",
+		ClientSecret:               "fakeone_a654+9sdf7+8we4f",
+		ApiKey:                     "",
+		Logger:                     zapLogger,
+		RetryMaxElapsedTimeSeconds: 300,
+	}
+
+	// authentication using API Key
+	authParamsKey = &AuthenticationParametersObj{
+		HTTPClient:                 *httpClientObj,
+		BackoffDefinition:          backoffDefinition,
+		EndpointURL:                "https://fake.api.com:443/BeyondTrust/api/public/v3/",
+		APIVersion:                 apiVersion,
+		ClientID:                   "",
+		ClientSecret:               "",
+		ApiKey:                     "fakeone_a654+9sdf7+8we4f",
+		Logger:                     zapLogger,
+		RetryMaxElapsedTimeSeconds: 300,
+	}
+}
+
+func TestSignOut(t *testing.T) {
+
+	InitializeGlobalConfig()
 
 	testConfig := UserTestConfig{
 		name: "TestSignOut",
@@ -64,7 +103,13 @@ func TestSignOut(t *testing.T) {
 		response: nil,
 	}
 
+<<<<<<< HEAD
 	var authenticate, _ = Authenticate(*httpClientObj, backoffDefinition, testConfig.server.URL, apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = Authenticate(*authParamsOauth)
+	apiUrl, _ := url.Parse(testConfig.server.URL + "/")
+	authenticate.ApiUrl = *apiUrl
+>>>>>>> main
 
 	err := authenticate.SignOut()
 	if err != nil {
@@ -73,17 +118,19 @@ func TestSignOut(t *testing.T) {
 }
 
 func TestSignAppin(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = Authenticate(*authParamsOauth)
+>>>>>>> main
 	testConfig := UserTestConfig{
 		name: "TestSignAppin",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -110,6 +157,7 @@ func TestSignAppin(t *testing.T) {
 }
 
 func TestSignAppinWithApiKey(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -121,6 +169,12 @@ func TestSignAppinWithApiKey(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = AuthenticateUsingApiKey(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, zapLogger, 300, "fake_api_key_")
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = AuthenticateUsingApiKey(*authParamsKey)
+>>>>>>> main
 	testConfig := UserTestConfig{
 		name: "TestSignAppinWithApiKey",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -147,17 +201,19 @@ func TestSignAppinWithApiKey(t *testing.T) {
 }
 
 func TestGetToken(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = Authenticate(*authParamsOauth)
+>>>>>>> main
 	testConfig := GetTokenConfig{
 		name: "TestGetToken",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -189,17 +245,19 @@ func TestGetToken(t *testing.T) {
 }
 
 func TestGetPasswordSafeAuthentication(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = Authenticate(*authParamsOauth)
+>>>>>>> main
 	testConfig := GetPasswordSafeAuthenticationConfig{
 		name: "TestGetPasswordSafeAuthentication",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {