+++ api/managed_account/managed_account_test.go
@@ -43,19 +43,46 @@ type CreateManagedAccountsResponse struct {
 // the default API version 3.0 will be used
 var apiVersion string = "3.1"
 
+<<<<<<< HEAD
 func TestManagedAccountGet(t *testing.T) {
+=======
+var authParams *authentication.AuthenticationParametersObj
+var zapLogger *logging.ZapLogger
+
+func InitializeGlobalConfig() {
+>>>>>>> main
 
 	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	zapLogger = logging.NewZapLogger(logger)
 
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
+<<<<<<< HEAD
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	authParams = &authentication.AuthenticationParametersObj{
+		HTTPClient:                 *httpClientObj,
+		BackoffDefinition:          backoffDefinition,
+		EndpointURL:                "https://fake.api.com:443/BeyondTrust/api/public/v3/",
+		APIVersion:                 apiVersion,
+		ClientID:                   "fakeone_a654+9sdf7+8we4f",
+		ClientSecret:               "fakeone_a654+9sdf7+8we4f",
+		ApiKey:                     "",
+		Logger:                     zapLogger,
+		RetryMaxElapsedTimeSeconds: 300,
+	}
+}
+
+func TestManagedAccountGet(t *testing.T) {
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 
 	testConfig := ManagedAccountTestConfig{
 		name: "TestManagedAccountGet",
@@ -87,8 +114,8 @@ func TestManagedAccountGet(t *testing.T) {
 }
 
 func TestManagedAccountCreateRequest(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -98,6 +125,9 @@ func TestManagedAccountCreateRequest(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountCreateRequest",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -124,8 +154,8 @@ func TestManagedAccountCreateRequest(t *testing.T) {
 }
 
 func TestCredentialByRequestId(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -135,6 +165,9 @@ func TestCredentialByRequestId(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestCredentialByRequestId",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -161,8 +194,8 @@ func TestCredentialByRequestId(t *testing.T) {
 }
 
 func TestManagedAccountRequestCheckIn(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -172,6 +205,9 @@ func TestManagedAccountRequestCheckIn(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountRequestCheckIn",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -198,8 +234,8 @@ func TestManagedAccountRequestCheckIn(t *testing.T) {
 }
 
 func TestManageAccountFlow(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -209,6 +245,9 @@ func TestManageAccountFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManageAccountFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -275,8 +314,8 @@ func TestManageAccountFlow(t *testing.T) {
 }
 
 func TestManageAccountFlowNotFound(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -286,6 +325,9 @@ func TestManageAccountFlowNotFound(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManageAccountFlowNotFound",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -352,8 +394,8 @@ func TestManageAccountFlowNotFound(t *testing.T) {
 }
 
 func TestSecretGetSecret(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -363,6 +405,9 @@ func TestSecretGetSecret(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestSecretGetSecret",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -427,8 +472,8 @@ func TestSecretGetSecret(t *testing.T) {
 }
 
 func TestSecretGetSecrets(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -438,6 +483,9 @@ func TestSecretGetSecrets(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestSecretGetSecrets",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -503,8 +551,8 @@ func TestSecretGetSecrets(t *testing.T) {
 }
 
 func TestManagedAccountFlowTechnicalErrorCreatingRequest(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -514,6 +562,9 @@ func TestManagedAccountFlowTechnicalErrorCreatingRequest(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowTechnicalErrorCreatingRequest",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -578,8 +629,8 @@ func TestManagedAccountFlowTechnicalErrorCreatingRequest(t *testing.T) {
 }
 
 func TestManagedAccountFlowBusinesslErrorCreatingRequest(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -589,6 +640,9 @@ func TestManagedAccountFlowBusinesslErrorCreatingRequest(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowBusinesslErrorCreatingRequest",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -653,8 +707,8 @@ func TestManagedAccountFlowBusinesslErrorCreatingRequest(t *testing.T) {
 }
 
 func TestManagedAccountFlowTechnicalErrorCredentialByRequestId(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -664,6 +718,9 @@ func TestManagedAccountFlowTechnicalErrorCredentialByRequestId(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowTechnicalErrorCredentialByRequestId",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -728,8 +785,8 @@ func TestManagedAccountFlowTechnicalErrorCredentialByRequestId(t *testing.T) {
 }
 
 func TestManagedAccountFlowBusinessErrorCredentialByRequestId(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -739,6 +796,9 @@ func TestManagedAccountFlowBusinessErrorCredentialByRequestId(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowBusinessErrorCredentialByRequestId",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -803,8 +863,8 @@ func TestManagedAccountFlowBusinessErrorCredentialByRequestId(t *testing.T) {
 }
 
 func TestManagedAccountFlowBusinessErrorAccountRequestCheckIn(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -814,6 +874,9 @@ func TestManagedAccountFlowBusinessErrorAccountRequestCheckIn(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowBusinessErrorAccountRequestCheckIn",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -878,8 +941,8 @@ func TestManagedAccountFlowBusinessErrorAccountRequestCheckIn(t *testing.T) {
 }
 
 func TestManagedAccountFlowTechnicalErrorAccountRequestCheckIn(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -889,6 +952,9 @@ func TestManagedAccountFlowTechnicalErrorAccountRequestCheckIn(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowTechnicalErrorAccountRequestCheckIn",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -953,8 +1019,8 @@ func TestManagedAccountFlowTechnicalErrorAccountRequestCheckIn(t *testing.T) {
 }
 
 func TestManagedAccountFlowGetAccountTechnicalError(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -964,6 +1030,9 @@ func TestManagedAccountFlowGetAccountTechnicalError(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountFlowGetAccountTechnicalError",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1009,8 +1078,8 @@ func TestManagedAccountFlowGetAccountTechnicalError(t *testing.T) {
 }
 
 func TestManageAccountFlowGetAccountBadResponse(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
+<<<<<<< HEAD
 	// create a zap logger wrapper
 	zapLogger := logging.NewZapLogger(logger)
 
@@ -1020,6 +1089,9 @@ func TestManageAccountFlowGetAccountBadResponse(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManageAccountFlowGetAccountBadResponse",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1064,6 +1136,7 @@ func TestManageAccountFlowGetAccountBadResponse(t *testing.T) {
 }
 
 func TestManagedAccountCreateManagedAccount(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1075,6 +1148,10 @@ func TestManagedAccountCreateManagedAccount(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := CreateManagedAccountsResponse{
 		name: "TestManagedAccountCreateManagedAccount",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1126,6 +1203,7 @@ func TestManagedAccountCreateManagedAccount(t *testing.T) {
 }
 
 func TestManagedAccountCreateManagedAccountExistingOne(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1137,6 +1215,10 @@ func TestManagedAccountCreateManagedAccountExistingOne(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountCreateManagedAccountExistingOne",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1179,6 +1261,7 @@ func TestManagedAccountCreateManagedAccountExistingOne(t *testing.T) {
 }
 
 func TestManagedAccountCreateManagedAccountFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1190,6 +1273,10 @@ func TestManagedAccountCreateManagedAccountFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := CreateManagedAccountsResponse{
 		name: "TestManagedAccountCreateManagedAccountFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1245,6 +1332,7 @@ func TestManagedAccountCreateManagedAccountFlow(t *testing.T) {
 }
 
 func TestManagedAccountCreateManagedAccountFlowSystemNotFound(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1256,6 +1344,10 @@ func TestManagedAccountCreateManagedAccountFlowSystemNotFound(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountCreateManagedAccountFlowSystemNotFound",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1304,6 +1396,7 @@ func TestManagedAccountCreateManagedAccountFlowSystemNotFound(t *testing.T) {
 }
 
 func TestManagedAccountCreateManagedAccountFlowEmptySystemList(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1315,6 +1408,10 @@ func TestManagedAccountCreateManagedAccountFlowEmptySystemList(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := ManagedAccountTestConfigStringResponse{
 		name: "TestManagedAccountCreateManagedAccountFlowEmptySystemList",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {