+++ api/secrets/secrets_test.go
@@ -37,18 +37,46 @@ type SecretTestConfigStringResponse struct {
 // the default API version 3.0 will be used
 var apiVersion string = "3.1"
 
+<<<<<<< HEAD
 func TestSecretGetSecretByPath(t *testing.T) {
+=======
+var authParams *authentication.AuthenticationParametersObj
+var zapLogger *logging.ZapLogger
+
+func InitializeGlobalConfig() {
+
+>>>>>>> main
 	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	zapLogger = logging.NewZapLogger(logger)
 
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
+<<<<<<< HEAD
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	authParams = &authentication.AuthenticationParametersObj{
+		HTTPClient:                 *httpClientObj,
+		BackoffDefinition:          backoffDefinition,
+		EndpointURL:                "https://fake.api.com:443/BeyondTrust/api/public/v3/",
+		APIVersion:                 apiVersion,
+		ClientID:                   "fakeone_a654+9sdf7+8we4f",
+		ClientSecret:               "fakeone_a654+9sdf7+8we4f",
+		ApiKey:                     "",
+		Logger:                     zapLogger,
+		RetryMaxElapsedTimeSeconds: 300,
+	}
+}
+
+func TestSecretGetSecretByPath(t *testing.T) {
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfig{
 		name: "TestSecretGetSecretByPath",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -80,16 +108,17 @@ func TestSecretGetSecretByPath(t *testing.T) {
 }
 
 func TestSecretGetFileSecret(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
 
-	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfig{
 		name: "TestSecretGetFileSecret",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -115,17 +144,19 @@ func TestSecretGetFileSecret(t *testing.T) {
 }
 
 func TestSecretFlow(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -179,17 +210,19 @@ func TestSecretFlow(t *testing.T) {
 }
 
 func TestSecretFlow_SecretNotFound(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlow_SecretNotFound",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -234,17 +267,19 @@ func TestSecretFlow_SecretNotFound(t *testing.T) {
 }
 
 func TestSecretGetSecret(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretGetSecret",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -273,17 +308,19 @@ func TestSecretGetSecret(t *testing.T) {
 }
 
 func TestSecretGetSecrets(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretGetSecrets",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -312,17 +349,19 @@ func TestSecretGetSecrets(t *testing.T) {
 }
 
 func TestSecretFlowTechnicalErrorFile(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowTechnicalErrorFile",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -373,17 +412,19 @@ func TestSecretFlowTechnicalErrorFile(t *testing.T) {
 }
 
 func TestSecretFlowBusinessErrorFile(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowBusinessErrorFile",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -434,17 +475,19 @@ func TestSecretFlowBusinessErrorFile(t *testing.T) {
 }
 
 func TestSecretFlowLongSecret(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowLongSecret",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -496,17 +539,19 @@ func TestSecretFlowLongSecret(t *testing.T) {
 }
 
 func TestSecretFlowTechnicalErrorCredential(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowTechnicalErrorCredential",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -551,17 +596,19 @@ func TestSecretFlowTechnicalErrorCredential(t *testing.T) {
 }
 
 func TestSecretFlowBusinessErrorCredential(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowBusinessErrorCredential",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -606,17 +653,19 @@ func TestSecretFlowBusinessErrorCredential(t *testing.T) {
 }
 
 func TestSecretFlowBadBody(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
 
-	// create a zap logger wrapper
-	zapLogger := logging.NewZapLogger(logger)
+	InitializeGlobalConfig()
 
+<<<<<<< HEAD
 	httpClientObj, _ := utils.GetHttpClient(5, false, "", "", zapLogger)
 
 	backoffDefinition := backoff.NewExponentialBackOff()
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFlowBadBody",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -660,6 +709,7 @@ func TestSecretFlowBadBody(t *testing.T) {
 }
 
 func TestSecretCreateTextSecretFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -671,6 +721,12 @@ func TestSecretCreateTextSecretFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateTextSecretFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -732,6 +788,7 @@ func TestSecretCreateTextSecretFlow(t *testing.T) {
 }
 
 func TestSecretCreateCredentialSecretFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -743,6 +800,12 @@ func TestSecretCreateCredentialSecretFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateCredentialSecretFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -804,6 +867,7 @@ func TestSecretCreateCredentialSecretFlow(t *testing.T) {
 }
 
 func TestSecretCreateFileSecretFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -815,6 +879,12 @@ func TestSecretCreateFileSecretFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateFileSecretFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -876,6 +946,7 @@ func TestSecretCreateFileSecretFlow(t *testing.T) {
 }
 
 func TestSecretCreateFileSecretFlowError(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -887,6 +958,12 @@ func TestSecretCreateFileSecretFlowError(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateFileSecretFlowError",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -946,6 +1023,7 @@ func TestSecretCreateFileSecretFlowError(t *testing.T) {
 }
 
 func TestSecretCreateBadInput(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -957,6 +1035,12 @@ func TestSecretCreateBadInput(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateBadInput",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1011,6 +1095,7 @@ func TestSecretCreateBadInput(t *testing.T) {
 }
 
 func TestSecretCreateSecretFlowFolderNotFound(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1022,6 +1107,12 @@ func TestSecretCreateSecretFlowFolderNotFound(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateSecretFlowFolderNotFound",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1070,6 +1161,7 @@ func TestSecretCreateSecretFlowFolderNotFound(t *testing.T) {
 }
 
 func TestSecretCreateSecretFlowEmptyFolderList(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1081,6 +1173,12 @@ func TestSecretCreateSecretFlowEmptyFolderList(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretCreateSecretFlowEmptyFolderList",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1129,6 +1227,7 @@ func TestSecretCreateSecretFlowEmptyFolderList(t *testing.T) {
 }
 
 func TestSecretFolderFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1140,6 +1239,12 @@ func TestSecretFolderFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFolderFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1186,6 +1291,7 @@ func TestSecretFolderFlow(t *testing.T) {
 }
 
 func TestSecretFolderFlowBadParentFolder(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1197,6 +1303,12 @@ func TestSecretFolderFlowBadParentFolder(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretFolderFlowBadParentFolder",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -1243,6 +1355,7 @@ func TestSecretFolderFlowBadParentFolder(t *testing.T) {
 }
 
 func TestSecretFolderFlowEmptyParentFolder(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1254,6 +1367,12 @@ func TestSecretFolderFlowEmptyParentFolder(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 
 	secretObj, _ := NewSecretObj(*authenticate, zapLogger, 4000)
 
@@ -1276,6 +1395,7 @@ func TestSecretFolderFlowEmptyParentFolder(t *testing.T) {
 }
 
 func TestSecretSafeFlow(t *testing.T) {
+<<<<<<< HEAD
 	logger, _ := zap.NewDevelopment()
 
 	// create a zap logger wrapper
@@ -1287,6 +1407,12 @@ func TestSecretSafeFlow(t *testing.T) {
 	backoffDefinition.MaxElapsedTime = time.Second
 
 	var authenticate, _ = authentication.Authenticate(*httpClientObj, backoffDefinition, "https://fake.api.com:443/BeyondTrust/api/public/v3/", apiVersion, "fakeone_a654+9sdf7+8we4f", "fakeone_aasd156465sfdef", zapLogger, 300)
+=======
+
+	InitializeGlobalConfig()
+
+	var authenticate, _ = authentication.Authenticate(*authParams)
+>>>>>>> main
 	testConfig := SecretTestConfigStringResponse{
 		name: "TestSecretSafeFlow",
 		server: httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {